Basic menus, game data and other improvements (#1)
Adds basic menus to the game, as well as game data (gmmain.nut), which creates a new game instance, as well as deletes it when quitGame() is called (currently done by pressing the Pause key in-game). Other fixes and code style improvements are also included, as well as the ability to toggle fullscreen mode, using F12.